example added in bash, plotting put in visualization
